fix ts issues from rebase
